fix(search/seed): fix inconsistencies that was happening between seeding function and syncing to search collection which ultimately lead to both failing.
This commit finally fixes the search collection.

+++ src/app/next/data/media-reset/script.ts
@@ -10,16 +10,38 @@ export async function clearMediaScript({
 }): Promise<void> {
   payload.logger.info(`↪ Resetting Media Collection...`)
 
-  await payload.db.deleteMany({
-    collection: 'media',
-    req,
-    where: {},
-  })
-  await payload.db.deleteVersions({
+  // Check if the media collection exists
+  const mediaCollection = await payload.find({
     collection: 'media',
     req,
     where: {},
   })
 
-  payload.logger.info('✓ Successfully reset Media Collection and Versions')
+  if (
+    !mediaCollection ||
+    !mediaCollection.docs ||
+    mediaCollection.docs.length === 0
+  ) {
+    payload.logger.info('✓ Media Collection is empty')
+    return
+  } else {
+    try {
+      payload.logger.info('Clearing Media Collection...')
+      await payload.db.deleteMany({
+        collection: 'media',
+        req,
+        where: {},
+      })
+      await payload.db.deleteVersions({
+        collection: 'media',
+        req,
+        where: {},
+      })
+
+      payload.logger.info('✓ Successfully reset Media Collection and Versions')
+    } catch (error) {
+      payload.logger.info(`Script failed. - ${error}`)
+      throw error
+    }
+  }
 }
